About Me [GA - Internal]
A section telling us more about you, your experience, and your contact information.
A picture of you in the center of the page, and two additional images with captions (one placed in the top right of your page, and one place in the bottom left).
An image carousel, with at least 3 images, written in JavaScript. Feel free to use outside resources in addition to the Dash tutorial, and don't be afraid to experiment - have fun with it!